Refactoring cmake
