+++ english/CMakeLists.txt
@@ -1,3 +1,4 @@
 add_subdirectory(collections)
 add_subdirectory(records)
 add_subdirectory(validators)
+add_subdirectory(handlers)
