+++ english/collections/translation_collection.h
@@ -1,10 +1,7 @@
 #pragma once
 
-#include "library/mongo/wrappers.h"
 #include "sources/collections/collection.h"
-
 #include "english/records/translation_record.h"
-
 #include "util/generic/maybe.h"
 
 namespace NEnglish {
