+++ english/collections/user_collection.h
@@ -1,8 +1,6 @@
 #pragma once
 
-#include "library/mongo/wrappers.h"
 #include "sources/collections/collection.h"
-
 #include "english/records/user_record.h"
 
 namespace NEnglish {
