--- english/handlers/admin/CMakeLists.txt
@@ -1,8 +1,0 @@
-include_directories(
-  ../../..
-)
-
-add_library(english_handlers STATIC
-  sessions.h
-  translations.h
-)
