+++ english/handlers/admin/sessions.cpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "english/handlers/common.h"
+#include "english/collections/session_collection.h"
+
+namespace NEnglish {
+
+    void GetAdminSessionsHandler(TDataSource& dataSource, const httplib::Request& req, NJson::TJsonValue& response) {
+        RestGetHandler(dataSource.English.CollectionSession, req, response);
+    }
+}
