+++ english/handlers/admin/sessions.h
@@ -1,20 +1,11 @@
 #pragma once
 
-#include "english/handlers/common.h"
-
 #include "contrib/httplib/httplib.h"
 #include "contrib/json/json.h"
 
-#include "english/collections/session_collection.h"
-
 #include "sources/data_source/data_source.h"
-
-#include "util/generic/ctype.h"
-#include "util/generic/iostream.h"
 
 namespace NEnglish {
 
-    void GetAdminSessionsHandler(TDataSource& dataSource, const httplib::Request& req, NJson::TJsonValue& response) {
-        RestGetHandler(dataSource.English.CollectionSession, req, response);
-    }
+    void GetAdminSessionsHandler(TDataSource& dataSource, const httplib::Request& req, NJson::TJsonValue& response);
 }
