+++ english/handlers/pagination.h
@@ -1,8 +1,6 @@
 #pragma once
 
 #include "contrib/httplib/httplib.h"
-
-#include "util/generic/ctype.h"
 #include "util/generic/string.h"
 
 namespace NEnglish {
@@ -13,9 +11,6 @@ namespace NEnglish {
         int limit;
         int skip;
 
-        TPagination(const httplib::Request& req) {
-            limit = NType::ToInt(req.GetParamValue(SIZE_OF_PAGE_PARAM, "0"));
-            skip = NType::ToInt(req.GetParamValue(NUMBER_OF_PAGE_PARAM, "0")) * limit;
-        }
+        TPagination(const httplib::Request& req);
     };
 }
