+++ english/validators/CMakeLists.txt
@@ -1,14 +1,15 @@
 include_directories(
-  ../..
+    ../..
 )
 
 add_library(english_validators STATIC
-  validator_common_english.cpp
-  validator_login.cpp
-  validator_user.cpp
-  validator_translation.cpp
+    validator_common_english.cpp
+    validator_login.cpp
+    validator_user.cpp
+    validator_translation.cpp
 )
 
 target_link_libraries(english_validators PUBLIC
-  english_collections
+    english_collections
+    validators
 )
