+++ sources/application/application.cpp
@@ -5,6 +5,7 @@
 #include "sources/handlers/hi.h"
 #include "sources/handlers/numbers.h"
 
+#include "english/handlers/common.h"
 #include "english/handlers/admin/sessions.h"
 #include "english/handlers/admin/translations.h"
 #include "english/handlers/users.h"
