+++ sources/data_source/CMakeLists.txt
@@ -1,13 +1,14 @@
 include_directories(
-  ../..
+    ../..
 )
 
 add_library(data_source STATIC
-  data_source.cpp
+    data_source.cpp
 )
 
 target_link_libraries(data_source PUBLIC
-  mongo
-  collections
-  english_collections
+    mongo
+    collections
+    english_collections
+    generic
 )
