+++ sources/records/CMakeLists.txt
@@ -6,3 +6,8 @@ add_library(records STATIC
     base_record.cpp
     record.cpp
 )
+
+target_link_libraries(records PUBLIC
+    json
+    generic
+)
