+++ sources/validators/CMakeLists.txt
@@ -1,7 +1,12 @@
 include_directories(
-  ../..
+    ../..
 )
 
 add_library(validators STATIC
     validator.cpp
+)
+
+target_link_libraries(validators PUBLIC
+    json
+    generic
 )
